Relocated vineyard endpoint unit tests
