+++ src/tests/tests.py
@@ -29,107 +29,6 @@ class MainTests(TestCase):
         response = client.get('/tests/')
         self.assertEqual(response.status_code, 200)
 
-    def test_response_vineyard_metadata(self):
-        setup_test_environment()
-        client = Client()
-        body = {
-            'vineyard_id': '0',
-            'auth_token': os.environ.get('LOGIN_SEC_TOKEN')
-        }
-        response = client.post(
-            '/vineyard',
-            data=json.dumps(body),
-            content_type='application/json'
-        )
-        self.assertEqual(response.status_code, 200)
-
-    @patch('cassy.get_vineyard_coordinates')
-    def test_response_vineyard_metadata_exception(self, vineyard_mock):
-        '''
-        Tests vineyard endpoint when cassy.get_vineyard_coordinates throws Exception
-        Args:
-            vineyard_mock:
-
-        Returns:
-
-        '''
-        setup_test_environment()
-        client = Client()
-        vineyard_mock.side_effect =Exception('Test exception')
-        body = {
-            'vineyard_id': '0',
-            'auth_token': os.environ.get('LOGIN_SEC_TOKEN')
-        }
-        response = client.post(
-            '/vineyard',
-            data=json.dumps(body),
-            content_type='application/json'
-        )
-        error = json.loads(response.content.decode('utf-8'))['errors']
-        self.assertTrue('vineyard_unknown' in error)
-        self.assertEqual(response.status_code, 400)
-
-    def test_response_vinemeta_invalid_vineyard(self):
-        setup_test_environment()
-        client = Client()
-        body = {
-            'vineyard_id': '-1',
-            'auth_token': os.environ.get('LOGIN_SEC_TOKEN')
-        }
-        response = client.post(
-            '/vineyard',
-            data=json.dumps(body),
-            content_type='application/json'
-        )
-        error = json.loads(response.content.decode('utf-8'))['errors']
-        self.assertTrue('vineyard_id_not_found' in error)
-        self.assertEqual(response.status_code, 400)
-
-    def test_response_vinemeta_vineyard_missing(self):
-        setup_test_environment()
-        client = Client()
-        body = {
-            'auth_token': os.environ.get('LOGIN_SEC_TOKEN')
-        }
-        response = client.post(
-            '/vineyard',
-            data=json.dumps(body),
-            content_type='application/json'
-        )
-        error = json.loads(response.content.decode('utf-8'))['errors']
-        self.assertTrue('vineyard_no_id' in error)
-        self.assertEqual(response.status_code, 400)
-
-    def test_response_vinemeta_invalid_vineyard_non_integer(self):
-        setup_test_environment()
-        client = Client()
-        body = {
-            'vineyard_id': 'abc',
-            'auth_token': os.environ.get('LOGIN_SEC_TOKEN')
-        }
-        response = client.post(
-            '/vineyard',
-            data=json.dumps(body),
-            content_type='application/json'
-        )
-        error = json.loads(response.content.decode('utf-8'))['errors']
-        self.assertTrue('vineyard_bad_id' in error)
-        self.assertEqual(response.status_code, 400)
-
-    def test_response_vineyard_metadata_invalid_token(self):
-        setup_test_environment()
-        client = Client()
-        body = {
-            'vineyard_id': '0',
-            'auth_token': 'chestercheetah'
-        }
-        response = client.post(
-            '/vineyard',
-            data=json.dumps(body),
-            content_type='application/json'
-        )
-        self.assertEqual(response.status_code, 403)
-
     def test_response_temperature_data(self):
         setup_test_environment()
         client = Client()
